requirements.txt
streamlit
torch
torchvision
numpy
pillow
ultralytics